requirements.txt
streamlit>=1.40.0
groq>=0.9.0
python-dotenv==1.0.1
